# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API",
    "version": "1.0.0",
    "description": "API for managing tasks and holiday tasks with OAuth (GitHub) authentication."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:3000/auth/github",  
            "tokenUrl": "http://localhost:3000/auth/github/callback", 
            "scopes": {
              "user": "User profile access"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/tasks/Tasks": {
      "get": {
        "summary": "Get all tasks",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "201": {
            "description": "Task created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/tasks/Tasks/{id}": {
      "put": {
        "summary": "Update a task",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/holidayTasks/HolidayTasks": {
      "get": {
        "summary": "Get all holiday tasks",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "200": {
            "description": "List of holiday tasks"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new holiday task",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "201": {
            "description": "Holiday task created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/holidayTasks/HolidayTasks/{id}": {
      "put": {
        "summary": "Update a holiday task",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "200": {
            "description": "Holiday task updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Delete a holiday task",
        "security": [
          {
            "bearerAuth": ["user"]
          }
        ],
        "responses": {
          "200": {
            "description": "Holiday task deleted"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/github": {
      "get": {
        "summary": "Authenticate via GitHub",
        "responses": {
          "200": {
            "description": "GitHub authentication initiated"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "summary": "GitHub OAuth callback",
        "responses": {
          "200": {
            "description": "GitHub OAuth callback received"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}